pyproject.toml
[project]
name = "mnist-1s-and-2s"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckdb>=1.2.1",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "tensorflow>=2.19.0",
]
